# Speed up muOS boot via background mounts, etc.