### Performance gains

I measured boot times in the `4d8da591` (White Room) hero build against the modified changes, using two approaches:

1. Total boot time (measured via video, from last frame before green LED to first frame after `muxlaunch` screen).
2. muOS startup time (measured via timestamps in `boot.log`).

I didn't do a statistically rigorous analysis, but FWIW, the time deltas in `boot.log` are consistent across several reboots.

| Device | Total boot time | muOS startup time | Logs |
| - | - | - | - |
| RG35XX-H (slow SD card) | 16 sec -> 12 sec | 8 sec -> 4 sec | [before](https://github.com/user-attachments/files/16823542/rg35xx.boot.base.log), [after](https://github.com/user-attachments/files/16823543/rg35xx.boot.modified.log) |
| RG40XX-H (faster SD card) | 12 sec -> 10 sec | 4 sec -> 2 sec | [before](https://github.com/user-attachments/files/16823544/rg40xx.boot.base.log), [after](https://github.com/user-attachments/files/16823545/rg40xx.boot.modified.log) |

See also the [videos used to compare total boot time](https://imgur.com/a/xMKndXa).

The speedup is more impactful on devices with slower storage, which makes sense, but in either case it appears to give about a 50% improvement to `boot.log` time, at least on my hardware and configuration.

### Changes

* (Main improvement) Starts mounting SD card storage partitions in the background as soon as udevd settles down, then continues booting in parallel, only blocking when storage bind mounts are absolutely necessary (e.g., we start touching the theme, catalogue, etc.). This is where most of the speedup comes from.
* Runs vmtouch preloading of RetroArch and its shared libs at idle IO priority so it'll use spare IOPS, but won't slow down more urgent parts of the startup sequence.
* Uses the default performance governor for the entirety of the boot process, only switching to ondemand when entering the frontend main loop.
* (Not performance related) Moves copying of `boot.log` to `/mnt/mmc/MUOS/log/boot` to right before the frontend main loop so that all the boot log messages appear there, including the ones from `frontend.sh`.

I've tested this on the RG35XX-H and RG40XX-H, and verified that SD2 bind mounts still work, as does SD1 fallback. I've also verified that hotplugging (and hot-unplugging) of SD2 still works (adjusting the bind mounts appropriately), as does hotplugging (and hot-unplugging) of external USB storage.